requirements.txt
click==6.6
Crowd==1.0.0
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.1
MarkupSafe==0.23
pyasn1==0.1.9
pycrypto==2.6.1
requests==2.11.0
Werkzeug==0.11.10
